% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_choices_for_question}
\alias{get_choices_for_question}
\title{Get choices for one question}
\usage{
get_choices_for_question(data, indicator, lg)
}
\arguments{
\item{data}{data of choices}

\item{indicator}{the name of the indicator for the question}

\item{lg}{language to use}
}
\value{
a data.frame to join
}
\description{
Get choices for one question
}
