% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R, R/names_of_sheet.R
\name{get_groups}
\alias{get_groups}
\title{Get groups form begin and end into a list with data and information}
\usage{
get_groups(data)

get_groups(data)
}
\arguments{
\item{data}{data from the survey sheet}
}
\value{
list

list
}
\description{
Get groups form begin and end into a list with data and information

Get groups form begin and end into a list with data and information
}
