% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\name{questionnaire}
\alias{questionnaire}
\title{questionnaire class is a class to load, check and manipulate one or more than one XLSForm}
\description{
questionnaire class is a class to load, check and manipulate one or more than one XLSForm

questionnaire class is a class to load, check and manipulate one or more than one XLSForm
}
\examples{

## ------------------------------------------------
## Method `questionnaire$new`
## ------------------------------------------------

ref <- questionnaire$new(
  path = system.file("household_survey_americas.xlsx", package = "surveyDesigner") 
)

head(ref$data$survey)

# Example by groups
ref$by_groups$group_intro
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-questionnaire-new}{\code{questionnaire$new()}}
\item \href{#method-questionnaire-clone}{\code{questionnaire$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-questionnaire-new"></a>}}
\if{latex}{\out{\hypertarget{method-questionnaire-new}{}}}
\subsection{Method \code{new()}}{
read the xlsx for each sheet and return a named list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{questionnaire$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path to the xlsForm}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
named list
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ref <- questionnaire$new(
  path = system.file("household_survey_americas.xlsx", package = "surveyDesigner") 
)

head(ref$data$survey)

# Example by groups
ref$by_groups$group_intro
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-questionnaire-clone"></a>}}
\if{latex}{\out{\hypertarget{method-questionnaire-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{questionnaire$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
