% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referential.R
\name{Referential}
\alias{Referential}
\title{Referential class is a class to load, check and manipulate the XLSForm}
\description{
Referential class is a class to load, check and manipulate the XLSForm

Referential class is a class to load, check and manipulate the XLSForm
}
\examples{

## ------------------------------------------------
## Method `Referential$new`
## ------------------------------------------------

ref <- Referential$new(
  path = system.file("SurveyDesigner_Referential.xlsx", package = "surveyDesigner") 
)

head(ref$data$survey)

# Example by groups
ref$by_groups$group_intro
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{named list for the referential file}

\item{\code{by_groups}}{survey modules separated by begin and end to manipulate data}

\item{\code{path}}{path for the xlsx file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Referential-new}{\code{Referential$new()}}
\item \href{#method-Referential-get_groups}{\code{Referential$get_groups()}}
\item \href{#method-Referential-clone}{\code{Referential$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Referential-new"></a>}}
\if{latex}{\out{\hypertarget{method-Referential-new}{}}}
\subsection{Method \code{new()}}{
read the xlsx for each sheet and return a named list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Referential$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path to the file with the full referential}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
named list
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ref <- Referential$new(
  path = system.file("SurveyDesigner_Referential.xlsx", package = "surveyDesigner") 
)

head(ref$data$survey)

# Example by groups
ref$by_groups$group_intro
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Referential-get_groups"></a>}}
\if{latex}{\out{\hypertarget{method-Referential-get_groups}{}}}
\subsection{Method \code{get_groups()}}{
get data by groups of begin and end
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Referential$get_groups()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Referential-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Referential-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Referential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
