% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_choices_for_question}
\alias{get_choices_for_question}
\title{Get choices for one question}
\usage{
get_choices_for_question(survey, full_name)
}
\arguments{
\item{survey}{data from the choices sheet}

\item{full_name}{the full name (i.e. concatenating groups) for the variable}
}
\value{
a data.frame to join
}
\description{
Get choices for one question
}
