% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Get groups form begin and end into a list with data and information}
\usage{
get_groups(data)
}
\arguments{
\item{data}{data form the survey sheet}
}
\value{
list
}
\description{
Get groups form begin and end into a list with data and information
}
